cad_server/requirements.txt
Django==3.0.2
djangorestframework==3.11.0
pymysql==0.9.3
requests==2.22.0